Deprecate convert object
